services/feed_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.0
geopy==2.4.0
httpx==0.27.0